ETL/python/object-storage/requirements.txt
boto3
python-dotenv
SQLAlchemy
awswrangler
mypy